Moves revalidation logic to `lib` (#1132)
